CsrfNamedGroup moved to other folder and namespace
